add enroll to app
